Fragment Communication

+++ app/src/main/java/com/edocent/employeeapp/data/EmployeeDetailFragment.java
@@ -1,34 +1,29 @@
 package com.edocent.employeeapp.data;
 
 
+import android.app.Fragment;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.edocent.employeeapp.R;
 
-/**
- * A simple {@link Fragment} subclass.
- */
 public class EmployeeDetailFragment extends Fragment {
 
-    int employeeIndex;
+    long employeeIndex;
 
-    public EmployeeDetailFragment() {
-        // Required empty public constructor
-    }
+    public EmployeeDetailFragment() { }
 
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
+
         return inflater.inflate(R.layout.fragment_employee_detail, container, false);
     }
 
-    public void setEmployeeIndex(int employeeIndex) {
+    public void setEmployeeIndex(long employeeIndex) {
         this.employeeIndex = employeeIndex;
     }
 }
